Api/requirements.txt
blinker>=1.0
certifi
charset-normalizer
click
colorama==0.4.6
Flask
idna
importlib-metadata
itsdangerous
Jinja2
MarkupSafe
newsapi-python==0.2.7
requests
urllib3
Werkzeug
zipp
